# Reduce spin-waiting in the thread pool on Arm processors