@kouvel 2025-05-08
Some perf numbers from Cobalt 100 below.

Json, 48-proc VM, 4096 connections

| Spin count | RPS       | Diff    | Procs used | Diff  | RPS per proc used | Diff  |
| -: | -: | -: | -: | -: | -: | -: |
| Before: 280 | 2,016,322 |         | 41.9       |       | 48145.8           |       |
| 140        | 2,014,717 | \-0.08% | 41.0       | \-0.9 | 49187.7           | 2.16% |
| After: 70  | 2,012,294 | \-0.20% | 40.1       | \-1.8 | 50177.8           | 4.22% |
| 35         | 2,003,038 | \-0.66% | 39.1       | \-2.8 | 51282.0           | 6.51% |
| 17         | 1,994,986 | \-1.06% | 39.0       | \-2.8 | 51099.4           | 6.13% |
| 8          | 1,975,699 | \-2.01% | 37.8       | \-4.1 | 52305.9           | 8.64% |
| 0          | 1,962,333 | \-2.68% | 38.3       | \-3.6 | 51216.1           | 6.38% |

- No regression in throughput, slight reduction in CPU usage
- Reducing the spin count further appears to regress throughput slightly

Orchard, 48-proc VM, 64 connections

| Spin count | RPS    | Diff    | Procs used | Diff  | RPS per proc used | Diff  |
| -: | -: | -: | -: | -: | -: | -: |
| Before: 280 | 29,806 |         | 43.2       |       | 690.0             |       |
| 140        | 29,595 | \-0.71% | 42.9       | \-0.3 | 690.5             | 0.07% |
| After: 70  | 29,275 | \-1.78% | 42.0       | \-1.2 | 697.5             | 1.09% |
| 35         | 28,582 | \-4.10% | 41.4       | \-1.8 | 691.2             | 0.18% |
| 17         | 28,089 | \-5.76% | 40.6       | \-2.6 | 691.6             | 0.24% |
| 0          | 26,910 | \-9.72% | 38.9       | \-4.3 | 692.4             | 0.36% |

- Small regression in throughput, slight reduction in CPU usage
- Reducing the spin count further appears to regress throughput more significantly

Orchard, 48-proc VM, 16 connections

| Spin count | RPS    | Diff    | Procs used | Diff  | RPS per proc used | Diff   |
| -: | -: | -: | -: | -: | -: | -: |
| Before: 280 | 11,442 |         | 14.9       |       | 767.3             |        |
| 140        | 11,484 | 0.36%   | 14.3       | \-0.7 | 805.2             | 4.94%  |
| After: 70  | 11,365 | \-0.68% | 13.4       | \-1.5 | 845.8             | 10.24% |
| 0          | 10,806 | \-5.56% | 12.2       | \-2.7 | 883.8             | 15.19% |

- No regression in throughput, small reduction in CPU usage. CPU efficiency (RPS per proc) improves with lower spin counts.
- Disabling spin-waiting appears to reduce throughput further with a limited number of connections, but with better CPU efficiency

Orchard, 48-proc VM, 8 connections

| Spin count | RPS     | Diff     | Procs used    | Diff  | RPS per proc used | Diff   |
| -: | -: | -: | -: | -: | -: | -: |
| Before: 280 | 6,044   |          | 7.8           |       | 771.9             |        |
| 140        | 5,954   | \-1.48%  | 7.2           | \-0.6 | 824.2             | 6.78%  |
| After: 70  | 5,997   | \-0.77%  | 6.9           | \-1.0 | 874.2             | 13.25% |
| 0          | 5,644   | \-6.61%  | 6.1           | \-1.8 | 930.9             | 20.60% |

- Similar to results for 16 connections, differences are a bit more pronounced

Orchard with Postgres DB on separate machine, 16-proc VM, 160 connections, 1600 fixed RPS

| Spin count | RPS     | Diff    | Procs used | Diff    | RPS per proc used | Diff   |
| -: | -: | -: | -: | -: | -: | -: |
| Before: 280 | 1,601.5 |         | 2.486      |         | 644.2             |        |
| 140        | 1,601.0 | \-0.03% | 2.217      | \-0.269 | 722.1             | 12.09% |
| After: 70  | 1,600.5 | \-0.06% | 2.033      | \-0.454 | 787.5             | 22.24% |
| 0          | 1,600.8 | \-0.05% | 1.862      | \-0.624 | 859.6             | 33.43% |

- Significant reduction in CPU usage and improvement in CPU efficiency

Orchard with Postgres DB on separate machine, 16-proc VM, 20 connections, 200 fixed RPS

| Spin count | RPS   | Diff    | Procs used | Diff    | RPS per proc used | Diff   |
| -: | -: | -: | -: | -: | -: | -: |
| Before: 280 | 200.3 |         | 0.421      |         | 475.9             |        |
| 140        | 200.0 | \-0.12% | 0.340      | \-0.081 | 588.7             | 23.69% |
| After: 70  | 200.0 | \-0.12% | 0.298      | \-0.123 | 672.3             | 41.25% |
| 0          | 200.0 | \-0.12% | 0.265      | \-0.156 | 754.7             | 58.58% |

- Similar to results for 160 connections and 1600 fixed RPS, differences are much more pronounced