@kouvel 2025-05-08
Some perf numbers from `aspnet-citrine-arm-lin` below.

Json, 80 procs, 256 connections

| Spin count | RPS     | Diff    | Procs used | Diff   | RPS per proc used | Diff   |
| -: | -: | -: | -: | -: | -: | -: |
| Before: 280 | 865,533 |         | 51.6       |        | 16761.6           |        |
| After: 70  | 856,088 | \-1.09% | 41.0       | \-10.6 | 20872.3           | 24.52% |
| 0          | 793,820 | \-8.29% | 35.2       | \-16.4 | 22558.1           | 34.58% |

Json, 80 procs, 512 connections

| Spin count | RPS       | Diff    | Procs used | Diff   | RPS per proc used | Diff   |
| -: | -: | -: | -: | -: | -: | -: |
| Before: 280 | 1,029,221 |         | 59.0       |        | 17431.9           |        |
| After: 70  | 1,029,566 | 0.03%   | 49.1       | \-10.0 | 20973.0           | 20.31% |
| 0          | 972,028   | \-5.56% | 43.1       | \-15.9 | 22540.6           | 29.31% |

Json, 80 procs, 1024 connections

| Spin count | RPS       | Diff    | Procs used | Diff   | RPS per proc used | Diff   |
| -: | -: | -: | -: | -: | -: | -: |
| Before: 280 | 1,113,428 |         | 62.4       |        | 17829.3           |        |
| After: 70  | 1,107,307 | \-0.55% | 55.4       | \-7.1  | 19994.9           | 12.15% |
| 0          | 1,091,217 | \-1.99% | 50.5       | \-12.0 | 21623.0           | 21.28% |

Fortunes, 80 procs, 256 connections

| Spin count | RPS     | Diff     | Procs used | Diff   | RPS per proc used | Diff   |
| -: | -: | -: | -: | -: | -: | -: |
| Before: 280 | 413,150 |          | 46.9       |        | 8799.8            |        |
| After: 70  | 390,563 | \-5.47%  | 36.8       | \-10.1 | 10601.7           | 20.48% |
| 0          | 346,682 | \-16.09% | 30.2       | \-16.7 | 11461.7           | 30.25% |

Fortunes, 80 procs, 512 connections

| Spin count | RPS     | Diff    | Procs used | Diff   | RPS per proc used | Diff   |
| -: | -: | -: | -: | -: | -: | -: |
| Before: 280 | 476,625 |         | 51.7       |        | 9217.9            |        |
| After: 70  | 471,417 | \-1.09% | 42.2       | \-9.5  | 11170.6           | 21.18% |
| 0          | 436,519 | \-8.41% | 35.2       | \-16.5 | 12387.6           | 34.39% |

Fortunes, 80 procs, 1024 connections

| Spin count | RPS     | Diff    | Procs used | Diff   | RPS per proc used | Diff   |
| -: | -: | -: | -: | -: | -: | -: |
| Before: 280 | 486,773 |         | 51.7       |        | 9414.6            |        |
| After: 70  | 497,833 | 2.27%   | 44.6       | \-7.1  | 11168.9           | 18.63% |
| 0          | 471,301 | \-3.18% | 37.2       | \-14.5 | 12663.2           | 34.51% |

- Small regression in throughput in Fortunes at 256 connections, though it's clear that the server can process higher RPS with more connections
- Significant reduction in CPU usage and improvement in CPU efficiency
- Disabling spin-waiting results in the best CPU efficiency, though with larger throughput regressions at limited lower connection counts